inlineにしたのと、warningを消した

+++ Game/client/graphic_2dui.cpp
@@ -1,7 +1,7 @@
 #include "graphic_2dui.hpp"
 
-const double PI = 3.141592;
-const double UI_MAGNIFICATION = 0.15/900 * Wd_Height;
+const float PI = 3.141592;
+const float UI_MAGNIFICATION = 0.15 / 900 * Wd_Height;
 
 void TimeImg::Set(TTF_Font *font){
     char numbers_string[10][2] = {"0","1", "2", "3", "4", "5", "6", "7", "8", "9"};
@@ -22,7 +22,7 @@ void TimeImg::DrawTime(int time, FloatRect dst){
         tmpdst = {dst.x + (dst.w*2/7)*i, dst.y, dst.w*2/7, dst.h};
         nums[tmpnum[i]].Draw(&tmpdst);
     }
-    tmpdst = {dst.x + dst.w/7 * 6.0, dst.y, dst.w/7, dst.h};
+    tmpdst = {(float)dst.x + (float)dst.w / 7 * 6.0f, (float)dst.y, (float)dst.w / 7, (float)dst.h};
     str_s.Draw(&tmpdst);
 
 }
@@ -38,8 +38,18 @@ void TimerUI::Set(TTF_Font *font){
     timer.Set(font);
     base.LoadImg(basefile);
     rotate.LoadImg(rotatefile);
-    dst = {Wd_Width*36/40 - UI_MAGNIFICATION * base.GetImgRect().w/2, Wd_Height * 5/30 - UI_MAGNIFICATION * base.GetImgRect().h/2, UI_MAGNIFICATION * base.GetImgRect().w, UI_MAGNIFICATION * base.GetImgRect().h};
-    timeDst = {Wd_Width*36/40 - UI_MAGNIFICATION * timeMagW * base.GetImgRect().w/2, Wd_Height * 16/90 - UI_MAGNIFICATION * timeMagH * base.GetImgRect().h/2, UI_MAGNIFICATION * timeMagW * base.GetImgRect().w, UI_MAGNIFICATION * timeMagH * base.GetImgRect().h};
+    dst = {
+            Wd_Width * 36 / 40 - UI_MAGNIFICATION * base.GetImgRect().w / 2,
+            Wd_Height * 5 / 30 - UI_MAGNIFICATION * base.GetImgRect().h / 2,
+            UI_MAGNIFICATION * base.GetImgRect().w,
+            UI_MAGNIFICATION * base.GetImgRect().h
+        };
+    timeDst = {
+            Wd_Width  * 36 / 40 - UI_MAGNIFICATION * (float)timeMagW * base.GetImgRect().w / 2,
+            Wd_Height * 16 / 90 - UI_MAGNIFICATION * (float)timeMagH * base.GetImgRect().h / 2,
+            UI_MAGNIFICATION * (float)timeMagW * base.GetImgRect().w,
+            UI_MAGNIFICATION * (float)timeMagH * base.GetImgRect().h
+        };
 }
 
 void TimerUI::Draw(float dir, float time){
@@ -63,9 +73,24 @@ void PlayerNameUI::Set(const PlayerData *pData, TTF_Font *font){
     base.LoadImg(basefile);
     name.LoadText(font,pData->name, color);
     ID.LoadText(font, idstr, color);
-    baseDst = {Wd_Width * 35/40 - UI_MAGNIFICATION * base.GetImgRect().w/2, Wd_Height * 24/30 - UI_MAGNIFICATION * base.GetImgRect().h/2, UI_MAGNIFICATION * base.GetImgRect().w, UI_MAGNIFICATION * base.GetImgRect().h};
-    nameDst = {Wd_Width * 71/80 - UI_MAGNIFICATION * nameMagW * base.GetImgRect().w/2, Wd_Height * 51/60 - UI_MAGNIFICATION * NameMagH * base.GetImgRect().h/2, UI_MAGNIFICATION * nameMagW * base.GetImgRect().w, UI_MAGNIFICATION * NameMagH * base.GetImgRect().h};
-    idDst = {Wd_Width * 307/320 - UI_MAGNIFICATION * idMagW * base.GetImgRect().w/2, Wd_Height * 87/120 - UI_MAGNIFICATION * idMagH * base.GetImgRect().h/2, UI_MAGNIFICATION * idMagW * base.GetImgRect().w, UI_MAGNIFICATION * idMagH * base.GetImgRect().h};
+    baseDst = {
+            Wd_Width * 35 / 40 - UI_MAGNIFICATION * base.GetImgRect().w / 2,
+            Wd_Height * 24 / 30 - UI_MAGNIFICATION * base.GetImgRect().h / 2,
+            UI_MAGNIFICATION * base.GetImgRect().w,
+            UI_MAGNIFICATION * base.GetImgRect().h
+        };
+    nameDst = {
+            Wd_Width * 71 / 80 - UI_MAGNIFICATION * (float)nameMagW * base.GetImgRect().w / 2,
+            Wd_Height * 51 / 60 - UI_MAGNIFICATION * (float)NameMagH * base.GetImgRect().h / 2,
+            UI_MAGNIFICATION * (float)nameMagW * base.GetImgRect().w,
+            UI_MAGNIFICATION * (float)NameMagH * base.GetImgRect().h
+        };
+    idDst = {
+            Wd_Width * 307 / 320 - UI_MAGNIFICATION * (float)idMagW * base.GetImgRect().w / 2,
+            Wd_Height * 87 / 120 - UI_MAGNIFICATION * (float)idMagH * base.GetImgRect().h / 2,
+            UI_MAGNIFICATION * (float)idMagW * base.GetImgRect().w,
+            UI_MAGNIFICATION * (float)idMagH * base.GetImgRect().h
+        };
 }
 
 void PlayerNameUI::Draw(){
@@ -86,13 +111,18 @@ void DirectionUI::Set(){
     Playerdir.LoadImg(dirfile);
     meterRect.LoadImg(meterfile);
 
-    dst = {Wd_Width * 4/40 - UI_MAGNIFICATION * base.GetImgRect().w/2, Wd_Height * 5/30 - UI_MAGNIFICATION * base.GetImgRect().h/2, UI_MAGNIFICATION * base.GetImgRect().w, UI_MAGNIFICATION * base.GetImgRect().h};
+    dst = {
+            Wd_Width * 4 / 40 - UI_MAGNIFICATION * base.GetImgRect().w / 2,
+            Wd_Height * 5 / 30 - UI_MAGNIFICATION * base.GetImgRect().h / 2,
+            UI_MAGNIFICATION * base.GetImgRect().w,
+            UI_MAGNIFICATION * base.GetImgRect().h
+        };
 }
 
 void DirectionUI::Draw(float playerdir, float rotdir){
 
-    FloatRect meterSrc = {0, 1-0.3*sin(5*rotdir)*sin(5*rotdir), 1, 1};
-    FloatRect meterDst = {dst.x, dst.y + dst.h*(1-0.3*sin(5*rotdir)*sin(5*rotdir)), dst.w, dst.h};
+    FloatRect meterSrc = {0, 1 - 0.3f * (float)sin(5 * rotdir) * (float)sin(5 * rotdir), 1, 1};
+    FloatRect meterDst = {dst.x, dst.y + dst.h * (1 - 0.3f * (float)sin(5 * rotdir) * (float)sin(5 * rotdir)), dst.w, dst.h};
     base.Draw(&dst);
     rotate.Draw(&dst,NULL,rotdir);
     Playerdir.Draw(&dst,NULL,playerdir);
@@ -222,7 +252,12 @@ void MiniMap::DestroyTexture(){
 void MapUI::Set(){
     base.LoadImg(basefile);
     minimap.SetMap();
-    dst = {Wd_Width * 4/40 - UI_MAGNIFICATION * base.GetImgRect().w/2, Wd_Height * 21/30 - UI_MAGNIFICATION * base.GetImgRect().h/2, UI_MAGNIFICATION * base.GetImgRect().w, UI_MAGNIFICATION * base.GetImgRect().h};
+    dst = {
+            Wd_Width * 4/40 - UI_MAGNIFICATION * base.GetImgRect().w/2,
+            Wd_Height * 21/30 - UI_MAGNIFICATION * base.GetImgRect().h/2,
+            UI_MAGNIFICATION * base.GetImgRect().w,
+            UI_MAGNIFICATION * base.GetImgRect().h
+        };
 }
 
 void MapUI::Draw(){
